OPNsense Firewall: Blocking IP's, Ports, and Protocols
In this project, I explored the OPNsense firewall, a tool crucial for network security. The focus was on understanding and implementing firewall rules and aliases to manage and secure networks effectively. The goal was to grasp practical insights into blocking specific IPs, protocols, and ports while optimizing organizational efficiency through alias usage.